🎨Refactor: Profile page divided business logic from component 🎨Product List: ADD porduct list , product listentry component
