+++ src/routes/ProductList/ProductList.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+import styled from "styled-components";
+import ProductListEntry from "../ProductListEntry/ProductListEntry";
+
+const Container = styled.ul`
+  display: flex;
+  gap: 10px;
+  overflow-x: scroll;
+`;
+
+export default function ProductList({ products }) {
+  return (
+    <Container>
+      {products.map((product) => {
+        return <ProductListEntry product={product} />;
+      })}
+    </Container>
+  );
+}
